Move HPC into its own module
Avoid triggering imports from other components
